+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 ### Fixed
 - Updated deprecated `typing` types to be compliant with Python >=3.11 guidelines (#216)
 - Fixed a typo in the badge for the OPI paper (#222).
+- Fixed `Structure.nelectrons` for structures containing ghost atoms (#268).
 
 ## [2.0.0] - 2026-02-10